refactoring

+++ include/romeo_moveit_actions/evaluation.hpp
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2016 SoftBank Robotics Europe
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+*/
+
 #ifndef EVALUATION_HPP
 #define EVALUATION_HPP
 
@@ -82,7 +98,7 @@ protected:
   std::string base_frame_;
 
   //step to test the working space
-  double test_step_;
+  float test_step_;
 
   //number of attempts
   int attempts_nbr_;
@@ -99,13 +115,18 @@ protected:
   //shift of the robot's base to teh floor
   double floor_to_base_height_;
 
-  //working space of the robot
-  double x_min_;
-  double x_max_;
-  double y_min_;
-  double y_max_;
-  double z_min_;
-  double z_max_;
+  /** working space in X dim min */
+  float x_min_;
+  /** working space in X dim max */
+  float x_max_;
+  /** working space in Y dim min */
+  float y_min_;
+  /** working space in Y dim max */
+  float y_max_;
+  /** working space in Z dim min */
+  float z_min_;
+  /** working space in Z dim max */
+  float z_max_;
 
   //default zero pose
   geometry_msgs::Pose pose_zero_;
